Don’t convert literals to String in test assertions
assert_eq and assert_ne works perfectly well when one argument is a String
and another is &str.  That is, it’s not necessary to convert the literal
argument to String for the test to compile and work correctly.

With that in mind, remove unnecessary `.to_string()` calls.  This makes
code easier to read (as it’s less ‘busy’).